[FIX] moved changeRole to Functions, fixed dropdown
